add Nuxt integration
