+++ packages/nuxt/playground/pages/test.vue
@@ -1,0 +1,5 @@
+<template>
+  <div>
+    <OnyxHeadline is="h1">Test page</OnyxHeadline>
+  </div>
+</template>
